migrate project to plain js
