+++ motoko/nft-creator/frontend/index.html
@@ -1,11 +1,14 @@
 <html lang='en'>
-    <head>
-        <meta charset='UTF-8' />
-        <meta name='viewport' content='width=device-width, initial-scale=1.0' />
-        <title>NFT Collection Management Application</title>
-    </head>
-    <body>
-        <div id='root'></div>
-        <script type='module' src='/src/main.tsx'></script>
-    </body>
+
+<head>
+    <meta charset='UTF-8' />
+    <meta name='viewport' content='width=device-width, initial-scale=1.0' />
+    <title>NFT Collection Management Application</title>
+</head>
+
+<body>
+    <div id='root'></div>
+    <script type='module' src='/src/main.jsx'></script>
+</body>
+
 </html>
