+++ motoko/nft-creator/frontend/src/components/AuthButton.jsx
