+++ motoko/nft-creator/frontend/src/components/CollectionClaim.jsx
