+++ motoko/nft-creator/frontend/src/components/OwnedNFTs.jsx
