+++ motoko/nft-creator/frontend/src/main.jsx
@@ -24,7 +24,7 @@ const queryClient = new QueryClient({
     },
 });
 
-ReactDOM.createRoot(document.getElementById("root")!).render(
+ReactDOM.createRoot(document.getElementById("root")).render(
     <QueryClientProvider client={queryClient}>
         <InternetIdentityProvider>
             <ToastProvider>
