--- motoko/nft-creator/frontend/src/types.ts
@@ -1,3 +1,0 @@
-import { Value__1 } from "declarations/backend/backend.did";
-
-export type Metadata = [] | [Array<[string, Value__1]>];
